feat: move to sveltekit (#71)
* feat: move to sveltekit

Replacing svelte-router based routing. This commit makes a minimal
set of changes required to get things functional. We'll follow up
with other changes to move things around, make them more
idiomatically sveltekit, and start leveraging some of its features
a little better.

* chore: minor version updates

* fix: refresh error

In cases where the sync completes immediately, the unsubscribe
callback isn't fully initialized. We were checking for this
logically, but the const declaration prevents access. var is more
permissive and defaults to undefined, so the && check works as
intended.

* chore: check in lockfile

* fix: rerun deductions on branch change

* fix: space detail tab navigation

* refactor: tab navigation on other show pages

* chore: normalize import paths

* ci: fix cypress tests

Moving to sveltekit complicates trying to intercept network calls,
as they can happen either on the client or server side. The approach
we're taking here is to serve the fixture data from the local preview
server during testing.

* chore: update vite middleware and imports