feat(vscode): share body parsing and external ref logic with core (#160)
* feat: initial working vscode web extension

* chore: update sveltekit and vite

This fixes the underlying issue that required a `skipLibCheck` in the first place.

* refactor: fold shared logic back in to core package

* fix: correct link in markdown hover preview

* fix: hover markdown links in codespaces

* docs: brain dump next steps

* chore: fix build-time warnings

* refactor: push Ref logic down to core

* chore: clean up imports

* refactor: simplify citation rendering

defining it as a function means we can simplify the reactivity

* refactor: push SerializedTheorem down to core

* feat: unify entity description parsing logic

and update the VSCode extension to use the body parsing logic in
hover previews. Still a few polishing items to do here:

- expand internal references to linked names
- fix katex styling oddities

but this is a reasonable enough start to at least publish as a
prerelease version.